Disable client_max_body_size checks in nginx
By default, nginx has a small (1MB) limit on the size of uploads,
which prevents using horizon interface to upload an image. Disabling
the client_max_body_size check allows for the glance configuration
to govern the maximum size image that should be accepted. This change
also disables proxy_buffering and proxy_request_buffering in order
to reduce buffering latency.

Closes-Bug: #1868503
Change-Id: I0a89e0845d6c7d2805556f87685d280b4e72122a