Fix snapcraft build issues
The latest version of snapcraft fails to build microstack. Chasing
it down is due to multiple staged parts which have conflicting
information. Fixing that bit causes the uwsgi services not to run
correctly.

This patch fixes the uwsgi services by no longer overriding the python
home directory, since python3 is staged into the snap we can use the
default python home.

This patch also removes the libc6 staging into the package, which should
generally be avoided.

Change-Id: I8c176689083831a0b8b56a192a9fbdfb50edbb66